fix: address PR review comments
- Only mark designs as processed after successful thumbnail regeneration,
  allowing failed/aborted designs to retry on next dialog open
- Extract duplicated calculateIdealDistance and camera constants into
  shared cameraFraming.ts utility, used by both thumbnail.ts and
  thumbnailRegenerator.ts
- Add abort signal check before expensive scene setup and rendering
  in thumbnailRegenerator to avoid unnecessary work when cancelled
